requirements.txt
# Core analysis
pandas>=2.0
numpy>=1.24
openpyxl>=3.1       # Excel file reading (NPCI + MOSPI)

# Dashboard generation
plotly>=5.18

# Scraper
undetected-chromedriver>=3.5
selenium>=4.15
requests>=2.31
python-dateutil>=2.8

# Notebook
jupyter>=1.0
nbformat>=5.9
